pyproject.toml
[tool.poetry]
name = "llama_summarizer"
version = "0.1.0"
description = "Built with llama"
authors = ["aygalic <aygalic@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.1"
llama-stack = "^0.0.38"
torch = "^2.4.1"
transformers = "^4.45.1"
accelerate = "^0.34.2"
llama-models = "^0.0.38"
protobuf = "^5.28.2"
gguf = "^0.10.0"
llama-cpp-python = "^0.3.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
